Move solver base classes to nektar_interface subdir.
